Added security policies

+++ app/Http/Controllers/CategoryController.php
@@ -37,9 +37,8 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-        // $this->authorize('create', Category::class);
         $request->validate([
-            'name' => 'required'
+            'name' => 'required|unique:categories'
         ]);
 
         $category = new Category;
@@ -80,9 +79,9 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
-        // $this->authorize('update', Category::class);
+        $this->authorize('update', $category);
         $request->validate([
-            'name' => 'required'
+            'name' => 'required|unique:categories'
         ]);
 
         $category->name = $request->name;
@@ -99,7 +98,7 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        // $this->authorize('delete', $category);
+        $this->authorize('delete', $category);
         $category->delete();
         return redirect()->action([CategoryController::class, 'index']);
     }
