+++ app/Http/Controllers/ClientController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\Http\Resources\ClientResource;
-use App\Models\Client;
 use App\Models\Order;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -11,6 +10,7 @@ use Illuminate\Http\Request;
 class ClientController extends Controller
 {
     private $prefix = 'intranet.clients.';
+
     /**
      * Display a listing of the resource.
      *
@@ -18,6 +18,7 @@ class ClientController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', User::class);
         return view($this->prefix . 'index', ['clients' => User::getClients()]);
     }
 
@@ -45,10 +46,10 @@ class ClientController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Client  $client
+     * @param  \App\Models\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function show(Client $client)
+    public function show(User $user)
     {
         //
     }
@@ -56,10 +57,10 @@ class ClientController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Client  $client
+     * @param  \App\Models\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function edit(Client $client)
+    public function edit(User $user)
     {
         //
     }
@@ -68,10 +69,10 @@ class ClientController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Client  $client
+     * @param  \App\Models\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Client $client)
+    public function update(Request $request, User $user)
     {
         //
     }
@@ -79,12 +80,12 @@ class ClientController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Client  $client
+     * @param  \App\Models\User $client
      * @return \Illuminate\Http\Response
      */
     public function destroy(User $client)
     {
-        // $this->authorize('delete', $client);
+        $this->authorize('delete', $client);
         $client->delete();
         return redirect()->action([ClientController::class, 'index']);
     }
